+++ examples/data-sources/forgejo_users/data-source.tf
@@ -1,0 +1,1 @@
+data "forgejo_users" "example" {}