+++ internal/provider/provider.go
@@ -79,5 +79,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 }
 
 func (p *Provider) DataSources(ctx context.Context) []func() datasource.DataSource {
-	return []func() datasource.DataSource{}
+	return []func() datasource.DataSource{
+		NewUsersDataSource,
+	}
 }