+++ CHANGELOG.md
@@ -6,4 +6,5 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Added provider configuration
+- Added provider configuration.
+- Added users data-source.